Refactor controller package structure and update FXML references
