--- src/main/java/edu/cairn/cis332finalproject/Scene3.java
@@ -1,4 +1,0 @@
-package edu.cairn.cis332finalproject;
-
-public class Scene3 {
-}
