+++ src/main/java/edu/cairn/cis332finalproject/controllers/LoginController.java
@@ -1,5 +1,6 @@
-package edu.cairn.cis332finalproject;
+package edu.cairn.cis332finalproject.controllers;
 
+import edu.cairn.cis332finalproject.Preload;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
