+++ src/main/java/edu/cairn/cis332finalproject/controllers/Scene2.java
@@ -1,0 +1,4 @@
+package edu.cairn.cis332finalproject.controllers;
+
+public class Scene2 {
+}
