+++ src/main/java/module-info.java
@@ -5,4 +5,6 @@ module edu.cairn.cis332finalproject {
 
     opens edu.cairn.cis332finalproject to javafx.fxml;
     exports edu.cairn.cis332finalproject;
+    exports edu.cairn.cis332finalproject.controllers;
+    opens edu.cairn.cis332finalproject.controllers to javafx.fxml;
 }
