+++ src/main/resources/edu/cairn/cis332finalproject/main.fxml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
-<AnchorPane xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.cairn.cis332finalproject.MainController">
+<AnchorPane xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.cairn.cis332finalproject.controllers.MainController">
    <fx:include source="navbar.fxml"/>
    <Label fx:id="data_field"
           text="Example Display in Center"
