+++ src/main/resources/edu/cairn/cis332finalproject/scene3.fxml
@@ -1,12 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
+<?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.*?>
 <AnchorPane xmlns="http://javafx.com/javafx/24.0.1"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="edu.cairn.cis332finalproject.Scene3"
+            fx:controller="edu.cairn.cis332finalproject.controllers.Scene3"
             prefHeight="400.0" prefWidth="600.0">
    <fx:include source="navbar.fxml"/>
    <Label text="Scene 3"
